# public/apidocs/swagger_v2.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.0.0",
    "title": "NOVELCovid/API",
    "description": "Covid19 Third Party API for Global Information",
    "license": {
      "name": "GNU V3",
      "url": "https://github.com/disease-sh/api/blob/master/LICENSE"
    }
  },
  "schemes": [
    "https",
    "http"
  ],
  "host": "disease.sh",
  "basePath": "/",
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v2/all": {
      "get": {
        "tags": [
          "Default"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get Global Totals for Actual and Yesterday Data",
        "description": "Get global stats: cases, deaths, recovered, time last updated, and active cases. Data is updated every 10 minutes",
        "responses": {
          "200": {
            "description": "Status OK"
          }
        }
      }
    },
    "/v2/states": {
      "get": {
        "tags": [
          "Default"
        ],
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "enum": [
              "cases",
              "todayCases",
              "deaths",
              "todayDeaths",
              "active"
            ],
            "description": "For this parameter you can provide a key from the country model (e.g. cases, todayCases, deaths, active, etc) to sort the states from greatest to least, depending on the key.",
            "type": "string"
          },
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          }
        ],
        "summary": "Get US State Totals",
        "description": "Get stats on United States of America States with COVID-19, including cases, new cases, deaths, new deaths, and active cases. Data is updated every 10 minutes",
        "responses": {
          "200": {
            "description": "Status OK"
          }
        }
      }
    },
    "/v2/states/{states}": {
      "get": {
        "tags": [
          "Default"
        ],
        "parameters": [
          {
            "name": "states",
            "in": "path",
            "required": true,
            "description": "state name or comma separated names spelled correctly",
            "type": "string"
          },
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          }
        ],
        "summary": "Get US State Totals for a specific state",
        "description": "Get cases, new cases, deaths, new deaths, and active cases for a state. Data is updated every 10 minutes",
        "responses": {
          "200": {
            "description": "Status OK"
          }
        }
      }
    },
    "/v2/continents": {
      "get": {
        "tags": [
          "Countries / Continents"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "enum": [
              "cases",
              "todayCases",
              "deaths",
              "todayDeaths",
              "recovered",
              "active",
              "critical",
              "casesPerOneMillion",
              "deathsPerOneMillion"
            ],
            "description": "For this parameter you can provide a key from the country model (e.g. cases, todayCases, deaths, recovered, active, etc) to sort the countries from greatest to least, depending on the key",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get All Continents Totals for Actual and Yesterday Data",
        "description": "Returns a JSON array with an element for each continent that has stats available. ",
        "responses": {
          "200": {
            "description": "Status OK",
            "schema": {
              "$ref": "#/definitions/continents"
            }
          }
        }
      }
    },
    "/v2/continents/{query}": {
      "get": {
        "tags": [
          "Countries / Continents"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterdays data",
            "type": "string"
          },
          {
            "name": "strict",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false"
            ],
            "description": "Defaults to true. Setting to false gives you the ability to fuzzy search countries. Example Oman vs. rOMANia",
            "type": "string"
          },
          {
            "name": "query",
            "in": "path",
            "required": true,
            "description": "The continent name",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get a Specific Continents Totals for Actual and Yesterday Data",
        "description": "Get the same data from the /v2/continents endpoint, but filter down to a specific continent.",
        "responses": {
          "200": {
            "description": "Status OK",
            "schema": {
              "$ref": "#/definitions/continent"
            }
          }
        }
      }
    },
    "/v2/countries": {
      "get": {
        "tags": [
          "Countries / Continents"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "enum": [
              "cases",
              "todayCases",
              "deaths",
              "todayDeaths",
              "recovered",
              "active",
              "critical",
              "casesPerOneMillion",
              "deathsPerOneMillion"
            ],
            "description": "For this parameter you can provide a key from the country model (e.g. cases, todayCases, deaths, recovered, active, etc) to sort the countries from greatest to least, depending on the key",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get All Countries Totals for Actual and Yesterday Data",
        "description": "Returns a JSON array with an element for each country that has stats available. This includes iso codes, lat/long, a link to the country flag, cases, new cases, deaths, new deaths, recovered, active cases, critical cases, and cases/deaths per one million people. Data is updated every 10 minutes",
        "responses": {
          "200": {
            "description": "Status OK",
            "schema": {
              "$ref": "#/definitions/countries"
            }
          }
        }
      }
    },
    "/v2/countries/{query}": {
      "get": {
        "tags": [
          "Countries / Continents"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          },
          {
            "name": "strict",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false"
            ],
            "description": "Defaults to true. Setting to false gives you the ability to fuzzy search countries. Example Oman vs. rOMANia",
            "type": "string"
          },
          {
            "name": "query",
            "in": "path",
            "required": true,
            "description": "For this parameter you can use Country Name && Country Id && ISOs (ISO 2 | ISO 3) 3166 Country Standards",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get a Specific Country's Totals for Actual and Yesterday Data",
        "description": "Get the same data from the /countries endpoint, but filter down to a specific country.",
        "responses": {
          "200": {
            "description": "Status OK",
            "schema": {
              "$ref": "#/definitions/country"
            }
          }
        }
      }
    },
    "/v2/countries/{country,country,...}": {
      "get": {
        "tags": [
          "Countries / Continents"
        ],
        "parameters": [
          {
            "name": "yesterday",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "For this parameter you can show yesterday data",
            "type": "string"
          },
          {
            "name": "query",
            "in": "path",
            "required": true,
            "description": "For this parameter you can use multiple Country Names && Country Ids && ISOs (ISO 2 | ISO 3) 3166 Country Standards separated by commas",
            "type": "string"
          },
          {
            "name": "allowNull",
            "in": "query",
            "required": false,
            "enum": [
              "true",
              "false",
              "1",
              "0"
            ],
            "description": "Lets you turn off the automatic null-to-0 transformation",
            "type": "string"
          }
        ],
        "summary": "Get a Specific Country's Totals for Actual and Yesterday Data",
        "description": "Get the same data from the /countries endpoint, but filter down to multiple specific countries. Data is updated every 10 minutes",
        "responses": {
          "200": {
            "description": "Status Ok"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "countryInfo": {
      "required": [
        "country",
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "number",
          "uniqueItems": true
        },
        "iso2": {
          "type": "string",
          "uniqueItems": true
        },
        "iso3": {
          "type": "string",
          "uniqueItems": true
        },
        "country": {
          "type": "string"
        },
        "lat": {
          "type": "number"
        },
        "long": {
          "type": "number"
        },
        "flag": {
          "type": "string"
        }
      }
    },
    "continent":  {
      "properties": {
        "continent": {
          "type": "string"
        },
        "updated": {
          "type": "number"
        },
        "cases": {
          "type": "number"
        },
        "todayCases": {
          "type": "number"
        },
        "deaths": {
          "type": "number"
        },
        "todayDeaths": {
          "type": "number"
        },
        "recovered": {
          "type": "number"
        },
        "active": {
          "type": "number"
        },
        "critical": {
          "type": "number"
        }
      }
    },
    "country": {
      "properties": {
        "country": {
          "type": "string"
        },
        "countryInfo": {
          "$ref": "#/definitions/countryInfo"
        },
        "cases": {
          "type": "number"
        },
        "todayCases": {
          "type": "number"
        },
        "deaths": {
          "type": "number"
        },
        "todayDeaths": {
          "type": "number"
        },
        "recovered": {
          "type": "number"
        },
        "active": {
          "type": "number"
        },
        "critical": {
          "type": "number"
        },
        "casesPerOneMillion": {
          "type": "number"
        },
        "deathsPerOneMillion": {
          "type": "number"
        },
        "updated": {
          "type": "number"
        }
      }
    },
...
    "continents": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/continent"
      }
    },
    "countries": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/country"
      }
    },
...
  },
...
}